refactor: settings.svelte to +page.svelte
